chevron sizes and rename files
