+++ frontend/src/assets/styles/_accordion.scss
@@ -2,8 +2,8 @@
   &::before {
     display: inline-block;
     content: '';
-    height: 10px;
-    width: 10px;
+    height: 8px;
+    width: 8px;
     margin-right: 12px;
     border-bottom: 2px solid currentColor;
     border-right: 2px solid currentColor;
